refactored the structure and branching logic
